fragment refactored
