+++ app/src/main/java/org/lniranjan/chatclone/ui/fragment/ChatListFragment.kt
@@ -1,0 +1,26 @@
+package org.lniranjan.chatclone.ui.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import org.lniranjan.chatclone.R
+
+
+class ChatListFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_chat_list, container, false)
+    }
+
+}
