+++ app/src/main/java/org/lniranjan/chatclone/ui/fragment/MessagingFragment.kt
@@ -1,0 +1,25 @@
+package org.lniranjan.chatclone.ui.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import org.lniranjan.chatclone.R
+
+
+class MessagingFragment : Fragment() {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+     }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_messaging, container, false)
+    }
+}
