+++ app/src/main/res/layout/fragment_contact_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactListFragment">
+    tools:context=".ui.fragment.ContactListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
